API/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
pymysql
PyJWT
python-multipart
cryptography